prottemps + makefile

+++ Practica2/ejercicio_prottemp_mejorado.c
@@ -9,12 +9,18 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <signal.h>
+#include <semaphore.h>
+#include <fcntl.h>
+#include <sys/stat.h>
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <unistd.h>
 #include <stdint.h>
 
+#define SEM1 "/sem1"
 int usr2 = 0;
+int N_global;
+int fin_trabajo = 0;
 
 /**
  * Nombre de la funcion: manejador
@@ -26,7 +32,6 @@ void manejador(int sig) {
     return;
 }
 
-
 /**
  * Nombre de la funcion: manejador_SIGUSR2
  * Parametros: sig: senal a manejar
@@ -34,10 +39,58 @@ void manejador(int sig) {
  * Return: 
  */
 void manejador_SIGUSR2(int sig) {
-    usr2++;
+    FILE *f = NULL;
+    char buf[100];
+    if ((f = fopen("data.txt", "r")) == NULL){
+        perror("fopen");
+        exit(EXIT_FAILURE);
+    }
+    fgets(buf, 100, f);
+    if (atoi(buf) == N_global){
+        fin_trabajo = 1;
+    }
     return;
 }
 
+void crear_txt(){
+    FILE *f = NULL;
+    if ((f = fopen("data.txt", "w")) == NULL){
+        perror("fopen");
+        exit(EXIT_FAILURE);
+    }
+    fprintf(f, "0\n");
+    fprintf(f, "0\n");
+    fclose(f);
+}
+
+/**
+ * Nombre de la funcion: trabajo_hijo
+ * Parametros: suma
+ * Descripcion: trabajo de lectura y escitura de un proceso hijo
+ * Return: 
+ */
+void trabajo_hijo(int suma){
+    FILE *f = NULL;
+    char buf[100];
+    int proc_done;
+    int sum_total;
+    if ((f = fopen("data.txt", "r")) == NULL){
+        perror("fopen");
+        exit(EXIT_FAILURE);
+    }
+    fgets(buf, 100, f);
+    proc_done = atoi(buf);
+    fgets(buf, 100, f);
+    sum_total = atoi(buf);
+    fclose(f);
+    if ((f = fopen("data.txt", "w")) == NULL){
+        perror("fopen");
+        exit(EXIT_FAILURE);
+    }
+    fprintf(f, "%d\n", proc_done+1);
+    fprintf(f, "%d\n", sum_total + suma);
+    fclose(f);
+}
 
 /**
  * Nombre de la funcion: main
@@ -55,6 +108,7 @@ int main(int argc, char* argv[]) {
     }
     
     int N = atoi(argv[1]);
+    N_global = N;
     int T = atoi(argv[2]);
     pid_t array_pid[N];
     pid_t pid;
@@ -62,6 +116,12 @@ int main(int argc, char* argv[]) {
     struct sigaction act;
     struct sigaction act_usr2;
     int i;
+
+    sem_t *sem1 = NULL;
+    if ((sem1 = sem_open(SEM1, O_CREAT | O_EXCL, S_IRUSR | S_IWUSR, 0)) == SEM_FAILED) {
+		perror("sem_open");
+		exit(EXIT_FAILURE);
+	}
 
     for (i = 0; i < N; i++){
         pid = fork();
@@ -97,6 +157,10 @@ int main(int argc, char* argv[]) {
             exit(EXIT_FAILURE);
         }
 
+        /* creacion data.txt */
+        crear_txt();
+        sem_post(sem1);
+
         if (alarm(T)) {
             fprintf(stderr, "Existe una alarma previa establecida\n");
         }
@@ -107,10 +171,12 @@ int main(int argc, char* argv[]) {
         sigsuspend(&set);
 
         /* enviar SIGTERM a todos los hijos */
-        for (i = 0; i < N; i++){
-            if (kill(array_pid[i], SIGTERM) < 0){
-                perror("kill");
-                exit(EXIT_FAILURE);
+        if (fin_trabajo){
+            for (i = 0; i < N; i++){
+                if (kill(array_pid[i], SIGTERM) < 0){
+                    perror("kill");
+                    exit(EXIT_FAILURE);
+                }
             }
         }
 
@@ -126,6 +192,13 @@ int main(int argc, char* argv[]) {
     int suma = 0;
     for (i = 1; i <= getpid()/10; i++) suma += i;
     fprintf(stdout, "PID = %jd\tSuma = %d\n",(intmax_t)getpid(), suma); 
+
+    /* espera a poder leer/escribir y entonces realiza tu trabajo */
+    sem_wait(sem1);
+    trabajo_hijo(suma); /* seccion critica */
+    sem_post(sem1);
+
+    /* el proceso ha finalizado su trabajo */
     if (kill(getppid(), SIGUSR2) < 0){
         perror("kill");
         exit(EXIT_FAILURE);
